requirements.txt
autopep8==1.4.4
Babel==2.7.0
beautifulsoup4==4.8.1
bootstrap4==0.1.0
dj-database-url==0.5.0
Django==2.2.7
django-bootstrap4==1.0.1
django-braces==1.13.0
django-heroku==0.3.1
django-model-utils==3.2.0
django-phonenumber-field==3.0.1
django-six==1.0.4
django-star-ratings==0.8.0
django-widget-tweaks==1.4.5
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
mccabe==0.6.1
phonenumberslite==8.11.0
Pillow==6.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
swapper==1.1.1
whitenoise==4.1.4
